requirements.txt
geopandas
python-igraph
pandas
